finally found the stupid bug making the navigation bar on www.kde.org disappear.
svn path=/trunk/kdelibs/; revision=48342

+++ khtml/rendering/bidi.cpp
@@ -287,7 +287,7 @@ public:
     unsigned short len;
     unsigned char level;
     unsigned short width;
-    short yOffset;
+    int yOffset;
 };
